refactor(web): rework dashboard into responsive app shell with htmx + Alpine (#156)
